Write code that executes an SQL file.
Successfully create the database by executing the file create_tables.sql. (Not certain I understand why it works this time as I had similar code that did not work before.) Delete prototypeTest.py as it will not be used. Change name of database in the sql file--I'll need to change it again later.