Fix Executor's constructor to work when database does not exist.
In the case where the project database has not yet been created, the constructor for Executor will now work, although, creating the project database and then re-instantiating the executor should be what you do immediately afterwards, as one cannot use most of the functionalities of Executor properly without connecting to the project database. Change connectToDatabase to be a non-static method. Change database back to "project" (in Executor and create_tables.sql) and change passwd to "password".