Verify that changes up to now run.
Change order of table definitions so that it actually runs. Comment some stuff out to test what I have--it works. (Partially) update the insertion example in executorDemo. Make Executor.insert() not check if tableName is valid--this allows for more freedom to specify which attributes are to be added for the tuple.